+++ sqle/driver/mysql/util/util_test.go
@@ -1,4 +1,4 @@
-package mysql
+package util
 
 import (
 	"testing"
